+++ src/sys/unix/mod.rs
@@ -54,6 +54,18 @@ cfg_os_poll! {
     #[cfg_attr(all(
         not(mio_unsupported_force_waker_pipe),
         any(
+            target_os = "android",
+            target_os = "espidf",
+            target_os = "fuchsia",
+            target_os = "hermit",
+            target_os = "illumos",
+            target_os = "linux",
+        )
+    ), path = "waker/eventfd.rs")]
+    #[cfg_attr(all(
+        not(mio_unsupported_force_waker_pipe),
+        not(mio_unsupported_force_poll_poll), // `kqueue(2)` based waker doesn't work with `poll(2)`.
+        any(
             target_os = "freebsd",
             target_os = "ios",
             target_os = "macos",
@@ -62,8 +74,34 @@ cfg_os_poll! {
             target_os = "watchos",
         )
     ), path = "waker/kqueue.rs")]
+    #[cfg_attr(any(
+        // NOTE: also add to the list list for the `pipe` module below.
+        mio_unsupported_force_waker_pipe,
+        all(
+            // `kqueue(2)` based waker doesn't work with `poll(2)`.
+            mio_unsupported_force_poll_poll,
+            any(
+                target_os = "freebsd",
+                target_os = "ios",
+                target_os = "macos",
+                target_os = "tvos",
+                target_os = "visionos",
+                target_os = "watchos",
+            ),
+        ),
+        target_os = "aix",
+        target_os = "dragonfly",
+        target_os = "haiku",
+        target_os = "netbsd",
+        target_os = "nto",
+        target_os = "openbsd",
+        target_os = "redox",
+        target_os = "solaris",
+        target_os = "vita",
+    ), path = "waker/pipe.rs")]
     mod waker;
-    pub(crate) use self::waker::Waker;
+    // NOTE: the `Waker` type is expected in the selector module as the
+    // `poll(2)` implementation needs to do some special stuff.
 
     mod sourcefd;
     #[cfg(feature = "os-ext")]
@@ -83,10 +121,22 @@ cfg_os_poll! {
         all(feature = "os-ext", not(target_os = "hermit")),
         // For the `Waker` type based on a pipe.
         mio_unsupported_force_waker_pipe,
+        all(
+            // `kqueue(2)` based waker doesn't work with `poll(2)`.
+            mio_unsupported_force_poll_poll,
+            any(
+                target_os = "freebsd",
+                target_os = "ios",
+                target_os = "macos",
+                target_os = "tvos",
+                target_os = "visionos",
+                target_os = "watchos",
+            ),
+        ),
+        // NOTE: also add to the list list for the `pipe` module below.
         target_os = "aix",
         target_os = "dragonfly",
         target_os = "haiku",
-        target_os = "illumos",
         target_os = "netbsd",
         target_os = "nto",
         target_os = "openbsd",
