+++ src/sys/unix/waker/pipe.rs
@@ -22,9 +22,10 @@ pub(crate) struct Waker {
 }
 
 impl Waker {
+    #[allow(dead_code)] // Not used by the `poll(2)` implementation.
     pub(crate) fn new(selector: &Selector, token: Token) -> io::Result<Waker> {
         let waker = Waker::new_unregistered()?;
-        selector.register(waker.fd.as_raw_fd(), token, Interest::READABLE)?;
+        selector.register(waker.receiver.as_raw_fd(), token, Interest::READABLE)?;
         Ok(waker)
     }
 
@@ -55,14 +56,7 @@ impl Waker {
         }
     }
 
-    #[cfg(any(
-        mio_unsupported_force_poll_poll,
-        target_os = "espidf",
-        target_os = "haiku",
-        target_os = "nto",
-        target_os = "solaris",
-        target_os = "vita",
-    ))]
+    #[allow(dead_code)] // Only used by the `poll(2)` implementation.
     pub(crate) fn ack_and_reset(&self) {
         self.empty();
     }
